+++ pkg/credentials/hex_test.go
@@ -1,0 +1,5 @@
+package credentials
+
+import "encoding/hex"
+
+func hexOf(key []byte) string { return hex.EncodeToString(key) }